Now using 32-bits size description, since Ted sends some huge messages that got truncated when using 16-bits description.

+++ include/CoSocket.h
@@ -71,7 +71,7 @@ public:
 	
 private:
 	int id;
-	quint16 blockSize;
+	quint32 blockSize;
 	string type;
 	
 private slots:
